Allow caller to define log level for listener
The current implementation always uses debug as log level. This makes
the logger less flexible as it requires the caller to use debug log
level for all logging to see the output.

This change allows the caller to set the desired log level in the
constructor. The default is still debug to retain current behavior.